+++ test/lib/Elastica/Test/LegacyScriptFieldsTest.php
@@ -1,0 +1,16 @@
+<?php
+namespace lib\Elastica\Test;
+
+use Elastica\ScriptFields as LegacyScriptFields;
+use Elastica\Test\Base as BaseTest;
+
+class LegacyScriptFieldsTest extends BaseTest
+{
+    /**
+     * @group unit
+     */
+    public function testParent()
+    {
+        $this->assertInstanceOf('Elastica\Script\ScriptFields', new LegacyScriptFields([]));
+    }
+}