+++ CHANGELOG.md
@@ -14,6 +14,7 @@ All notable changes to this project will be documented in this file based on the
 ### Improvements
 
 ### Deprecated
+- Elastica\AbstractScript|Script|ScriptFile|ScriptFields deprecated in favor of Elastica\Script|AbstractScript|Script|ScriptFile|ScriptFields [#1028](https://github.com/ruflin/Elastica/pull/1028)
 
 
 ## [3.0.1](https://github.com/ruflin/Elastica/compare/3.0.0...3.0.1)